createtestdb/requirements.txt
awscli==1.29.49
boto3==1.28.49
botocore==1.31.49
colorama==0.4.4
dnspython==2.4.2
docutils==0.16
jmespath==1.0.1
Pillow==10.0.1
pyasn1==0.5.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.6.2
six==1.16.0
text-to-image==0.0.5
urllib3==1.26.16
